+++ spec/sync/index_spec.js
@@ -10,12 +10,12 @@ var lazy = lazy_require({
 });
 
 describe("Azk sync, main module", function() {
-  var example_fixtures = h.fixture_path('sync/test_1/');
   var invalid_fixtures = path.join(h.fixture_path('sync/test_1/'), 'invalid');
 
-  function make_copy() {
+  function make_copy(fixture_path = 'sync/test_1/') {
+    var fixtures = h.fixture_path(fixture_path);
     return all([
-      h.copyToTmp(example_fixtures),
+      h.copyToTmp(fixtures),
       h.tmp_dir()
     ]);
   }
@@ -30,24 +30,25 @@ describe("Azk sync, main module", function() {
   it("should sync two folders", function() {
     return async(function* () {
       var [origin, dest] = yield make_copy();
-      var code   = yield lazy.Sync.sync(origin, dest);
-      var result = yield h.diff(origin, dest);
-      h.expect(code).to.eql(0);
-      h.expect(result).to.have.property('deviation', 0);
+      var result         = yield lazy.Sync.sync(origin, dest);
+      var diff           = yield h.diff(origin, dest);
+
+      h.expect(result).to.have.property('code', 0);
+      h.expect(diff).to.have.property('deviation', 0);
     });
   });
 
   it("should sync one file between two folders", function() {
     return async(function* () {
-      var include = "bar/clothes/barney.txt";
+      var include        = "bar/clothes/barney.txt";
       var [origin, dest] = yield make_copy();
-      var code   = yield lazy.Sync.sync(origin, dest, { include });
+      var result         = yield lazy.Sync.sync(origin, dest, { include });
 
       // Compare folders and files
       var result_folder = yield h.diff(origin, dest);
       var result_file   = yield h.diff(path.join(origin, include), path.join(dest, include));
 
-      h.expect(code).to.eql(0);
+      h.expect(result).to.have.property('code', 0);
       h.expect(result_file).to.have.property('deviation', 0);
       h.expect(result_folder).to.have.property('deviation', 10);
     });
@@ -55,15 +56,15 @@ describe("Azk sync, main module", function() {
 
   it("should sync two folders but exclude a file list", function() {
     return async(function* () {
-      var except = "foo/";
+      var except         = "foo/";
       var [origin, dest] = yield make_copy();
-      var code   = yield lazy.Sync.sync(origin, dest, { except });
+      var result         = yield lazy.Sync.sync(origin, dest, { except });
 
       // Compare folders and subfolders
       var result_folder    = yield h.diff(origin, dest);
       var result_subfolder = yield h.diff(path.join(origin, "bar"), path.join(dest, "bar"));
 
-      h.expect(code).to.eql(0);
+      h.expect(result).to.have.property('code', 0);
       h.expect(result_folder).to.have.property('deviation', 3);
       h.expect(result_subfolder).to.have.property('deviation', 0);
     });
@@ -71,17 +72,30 @@ describe("Azk sync, main module", function() {
 
   it("should sync two folders but exclude paths from text file", function() {
     return async(function* () {
-      var except_from = h.fixture_path("sync/rsyncignore.txt");
+      var except_from    = h.fixture_path("sync/rsyncignore.txt");
       var [origin, dest] = yield make_copy();
-      var code   = yield lazy.Sync.sync(origin, dest, { except_from });
+      var result         = yield lazy.Sync.sync(origin, dest, { except_from });
 
       // Compare folders and subfolders
       var result_folder    = yield h.diff(origin, dest);
       var result_subfolder = yield h.diff(path.join(origin, "foo"), path.join(dest, "foo"));
 
-      h.expect(code).to.eql(0);
+      h.expect(result).to.have.property('code', 0);
       h.expect(result_folder).to.have.property('deviation', 7);
       h.expect(result_subfolder).to.have.property('deviation', 0);
+    });
+  });
+
+  it("should sync two folders containing special character on path", function() {
+    return async(function* () {
+      var [origin, dest] = yield make_copy('test-app/special:\'` "\\');
+      dest = path.join(dest, 'special:\'` "\\');
+
+      var result = yield lazy.Sync.sync(origin, dest);
+      var diff   = yield h.diff(origin, dest);
+
+      h.expect(result).to.have.property('code', 0);
+      h.expect(diff).to.have.property('deviation', 0);
     });
   });
 
@@ -98,24 +112,27 @@ describe("Azk sync, main module", function() {
   h.describeRequireVm("with enabled vm", function() {
     it("should sync two folders", function() {
       return async(function* () {
-        var name = config("agent:vm:name");
-        var dest = path.join("/tmp", lazy.uuid.v4());
-        var opts = { ssh: lazy.Client.ssh_opts() };
+        var name = config('agent:vm:name');
+        var dest = path.join('/tmp', lazy.uuid.v4(), "a b'`\\\"");
+        var opts = { ssh: lazy.Client.ssh_opts(), except: ['test file 2'] };
+        var example_fixtures = h.fixture_path('test-app/special:\'` "\\');
 
         // Make destination folder
-        var vm_code = yield lazy.VM.ssh(name, "mkdir -p " + dest);
+        var vm_code = yield lazy.VM.ssh(name, "mkdir -p " + path.join(dest, ".."));
         h.expect(vm_code).to.equal(0);
 
         // Sync folders
-        var code = yield lazy.Sync.sync(example_fixtures, dest, opts);
-        h.expect(code).to.eql(0);
+        var result = yield lazy.Sync.sync(example_fixtures, dest, opts);
+        h.expect(result).to.have.property('code', 0);
 
         // Test destination folder in vm
-        var file    = path.join(dest, "bar/Fred.txt");
-        var folder  = path.join(dest, "bar/clothes");
-        var cmd     = "test -f " + file + " && test -d " + folder;
+        dest = dest.replace(/([`"\\])/g, '\\$1');
+        var file   = path.join(dest, 'test file 1');
+        var folder = path.join(dest, 'test file 2');
+        var cmd    = `test -f "${file}" && test ! -f "${folder}"`;
         vm_code = yield lazy.VM.ssh(name, cmd);
-        h.expect(vm_code).to.equal(0);
+
+        h.expect(vm_code).to.equal(0, 'files no synced to destination');
       });
     });
   });